+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support `InitVar`
 
+## Fixed 
+
+- Fix `Union` type hooks
+
 ## [1.3.0] - 2020-03-14
 
 ### Added